runtime: add wrappers for sys* functions and consolidate docs
This change lifts all non-platform-specific code out of sys* functions
for each platform up into wrappers, and moves documentation about the OS
virtual memory abstraction layer from malloc.go to mem.go, which
contains those wrappers.

Reviewed-on: https://go-review.googlesource.com/c/go/+/393398
Trust: Michael Knyszek <mknyszek@google.com>
